Add UserController with AddUser endpoint and DI setup
Introduced a new UserController in WebAPI.Controllers to handle
HTTP POST requests for adding new users. The controller uses
dependency injection to get an instance of IUserService and
defines an AddUser endpoint that takes an AddUserRequest object
containing user details. The input is validated and the user is
added via the AddUserAsync method of IUserService. Also added
AddUserRequest class to encapsulate user data. Registered
IUserService with its implementation UserService in Program.cs.